# algoAPI/swagger.yaml
openapi: 3.0.3
info:
  title: Algo API
  description: >-
    Documentation for Algo API. REST API service, offering algorithm execution
    on request. Additionaly, input data can be saved to the service and used
    again. Most paths require a user account, which is managed through the API
    aswell.
  version: '1.0'
  x-logo:
    url: ''
servers:
  - url: 'https://algo-algo.herokuapp.com/'
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Creates a user
      description: 'Used for creating a user account, returns a JWT token.'
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              username: test
              email: test@test.com
              password: test
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/registration_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
...
  '/data/{data_id}':
    get:
      tags:
        - data
      summary: Returns the requested data
      description: Used for fetching data.
      operationId: getData
      parameters:
        - name: data_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Data_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Not authorized to access data
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - JWTAuth: []
    put:
      tags:
        - data
      summary: Updates the data
      description: >-
        Used for updating data. User can update data and allowed field. Allowed
        is used for authorizing other users to use your data. Array has to
        include their user id. Id "000000000000000000000000" allows every user
        to access data.
      operationId: updateData
      parameters:
        - name: data_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/data_update'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Data_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Not authorized to access data
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Internal server error
      security:
        - JWTAuth: []
    delete:
      tags:
        - data
      summary: Deletes the data
      description: Used for deleting data.
      operationId: deleteData
      parameters:
        - name: data_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Not authorized to access data
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Internal server error
      security:
        - JWTAuth: []
...
components:
  schemas:
...
    Error:
      required:
        - message
      type: object
      properties:
        message:
          type: string
          description: Error message indicating the issue
      example:
        message: User not found
    User:
      required:
        - email
        - password
        - username
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        username: test
        email: test@test.com
        password: test
    Data_response:
      properties:
        id:
          type: string
        data:
          type: string
        user:
          type: string
        allowed:
          type: array
          items:
            type: string
        created:
          type: string
        updated:
          type: string
      example:
        id: 5f3348e14d9b386e025681a1
        data: '-3,6,1,7'
        user: 5f330a4df21982980282af97
        allowed:
          - '000000000000000000000000'
          - 5fe98a1453590b33f7a82d64
        created: '2020-12-20T03:57:51+01:00'
        updated: '2020-12-20T03:57:51+01:00'
...
    registration_response:
      required:
        - created
        - email
        - id
        - username
      properties:
        userID:
          type: string
        token:
          type: string
        username:
          type: string
      example:
        userID: 5fe98a1453590b33f7a82d64
        token: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MDkxNDQzNTYsInVzZXJfaWQiOiI1ZmU5OGExNDUzNTkwYjMzZjdhODJkNjQiLCJ1c2VyX3VzZXJuYW1lIjoidGVzdCJ9.jgLcRffdVVsOzdfGDl09pcQXruEnpolNuyatPp8CflM
        username: test
...
    data_update:
      properties:
        data:
          type: string
        allowed:
          type: array
          items:
            type: string
      example:
        data: '-3,6,1,7'
        allowed:
          - '000000000000000000000000'
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT